requirements.txt
rembg
Pillow
flask